+++ feincms3/plugins/image.py
@@ -3,6 +3,8 @@ Provides an image plugin with support for setting the primary point of
 interest. This is very useful especially when cropping images. Depends on
 `django-imagefield <https://django-imagefield.readthedocs.io>`__.  """
 
+import os
+
 from content_editor.admin import ContentEditorInline
 from django.db import models
 from django.utils.html import format_html
@@ -48,7 +50,7 @@ class Image(models.Model):
         verbose_name_plural = _("images")
 
     def __str__(self):
-        return self.image.name
+        return self.caption or os.path.basename(self.image.name)
 
 
 class ImageInline(ContentEditorInline):